refactor: MVC 패턴 구조로 변경
