+++ src/test/kotlin/racingcar/CarTest.kt
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.controller.validateWinnersCount
+import racingcar.domain.Car
 
 class CarTest {
     @Test
